+++ annotations/nocedit.exe/pseudocode/src/core/motion.cpp/CMotionController_jumpToMotion_FUN_0052dde0.cpp
@@ -17,13 +17,13 @@ core_motion_cpp_CMotionController_jumpToMotion_FUN_0052dde0
   
   if (target_frame_number == -1.0f) {
     iVar1 = (*this_ptr->vtable->findPatchToFrame)
-                      (this_ptr,this_ptr->current_motion_index,this_ptr->current_frame_number,
-                       target_motion_index);
+                      (this_ptr,this_ptr->current_motion_index,(float)this_ptr->current_frame_number
+                       ,target_motion_index);
     in_stack_0000001c = (float)iVar1;
   }
   this_ptr->in_transition = 0;
   this_ptr->current_motion_index = in_stack_00000018;
-  this_ptr->current_frame_number = in_stack_0000001c;
+  this_ptr->current_frame_number = (int)in_stack_0000001c;
   core_motion_cpp_CMotionController_clearTweenState_FUN_0052de40(this_ptr);
   return;
 }
