+++ annotations/nocedit.exe/pseudocode/src/core/motion.cpp/CMotionController_reverseTransition_FUN_0052da50.cpp
@@ -17,8 +17,8 @@ core_motion_cpp_CMotionController_reverseTransition_FUN_0052da50(CMotionControll
   iVar1 = this_ptr->current_motion_index;
   this_ptr->current_motion_index = this_ptr->tween_target_motion;
   this_ptr->tween_target_motion = iVar1;
-  fVar2 = this_ptr->current_frame_number;
-  this_ptr->current_frame_number = this_ptr->tween_target_frame;
+  fVar2 = (float)this_ptr->current_frame_number;
+  this_ptr->current_frame_number = (int)this_ptr->tween_target_frame;
   this_ptr->tween_target_frame = fVar2;
   if (this_ptr->tween_type == 5) {
     this_ptr->tween_type = 4;
