+++ annotations/nocedit.exe/pseudocode/src/core/motion.cpp/CMotionController_save_FUN_0052e670.cpp
@@ -20,7 +20,7 @@ core_motion_cpp_CMotionController_save_FUN_0052e670(CMotionController *this_ptr,
       g_CurrentLineNumber = 0x536;
       core_main_c_displayErrorAndQuit_FUN_00506f10("Tried to save motion controller state without setting motion list pointer!");
     }
-    dVar2 = (double)this_ptr->current_frame_number;
+    dVar2 = (double)(float)this_ptr->current_frame_number;
     pSVar1 = core_motion_cpp_CMotionController_getCurrentMotion_FUN_0052dab0(this_ptr);
     _fprintf(file_handle,"%s\"%s\",%g // current motion name, curFrameNumber\n",in_stack_0000000c,pSVar1,dVar2);
     return;
